Update Script - YouTube API Discovery URL Fix
For some reason in recent builds the API stopped working and couldn't find the YouTube API, but only on exe builds. Now the code specifies the google api discovery url directly which solves it.

Added exception catch for when token can't be refreshed, to re-authenticate.

Changed double to single quotes in api function for consistency