Update Script- Massive reply scanning speed improvement
Now fetches a limited number of replies along with their top/parent comments. This means a separate API call does not need to be made for threads with less than ~5 replies.

For videos with many replies, this speeds up scanning by as much as 50x !

Also improved speed of getting video titles when scanning entire channel.