Add ability to set custom configs folder
User can now set directory  in which the program will search for config files. It is set with a new config setting: configs_path. The default folder is 'configs'.

If user changes from the default, they must keep the main config file in the same folder as the program, and put secondary configs in their chosen directory