Fix Rare Crash from Deleted Channel when Writing JSON Log
When a comment from a deleted channel is found, it caused a mis-match between the length of expected info items from API and the actual returned number, causing an Index Error. Program now iterates through the response based on the response itself, instead of the expected input, in case there is a mismatch.

Addresses #672