Fix Bad Parameter Checking Comment Existence, Improve Error Message
Fixed unexpected parameter error when checking for success of deleted / recovered comments. The 'maxResults' parameter cannot be used when 'id' parameter is used.

Also fixed another resulting exception where 'results' variable was called, but wouldn't exist because the API did not return that value.

Also now when an HttpError occurs when attempting to check comments, it will print the reason along side the comment's info.