Add encryption of token file by default
Adds new encrypt_token_file option in config, which is True by default. If True, next time the program is run, or whenever the user logs in, the token file will be replaced by an encrypted version. The user will enter the password for this.

If False, it will not have the user enter a password to encrypt the file. But if an encrypted token file already exists, it will still prompt the user for the password.