fix highlight of TJoe.io links
Not a common reality on Windows shells but many
Linux shells have a feature that you can click,
or Ctrl+Click on links to open them but right
now the links to TJoe.io are incomplete and not
recognized as links by the shell.

I tested it on Kitty and now clicking on links
works!